--- payment-service/src/main/java/com/sobok/paymentservice/payment/controller/payment/PaymentController.java
@@ -1,26 +1,0 @@
-package com.sobok.paymentservice.payment.controller.payment;
-
-import com.sobok.paymentservice.common.dto.ApiResponse;
-import com.sobok.paymentservice.payment.dto.payment.PaymentRegisterReqDto;
-import com.sobok.paymentservice.payment.service.payment.PaymentService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/payment")
-@RequiredArgsConstructor
-@Slf4j
-public class PaymentController {
-    private final PaymentService paymentService;
-
-    @PostMapping("/register")
-    public ResponseEntity<?> registerPayment(@RequestBody PaymentRegisterReqDto reqDto) {
-        Long paymentId = paymentService.registerPayment(reqDto);
-        return ResponseEntity.ok().body(ApiResponse.ok(paymentId, "주문 사전 정보가 정상적으로 저장되었습니다."));
-    }
-}
